Uploaded initial files
LoadData.py file has an extract data function, which creates a datablock object (DataBlock.py) with the raw_data from files in the import directory. It creates a directory for files created and saves the unique words for both modules and programmes as well as the use of keywords (found in keywords.txt) in a text file. 

DataBlock.py file contains the OOP which stores the raw_data into programmes and modules, it creates an individual directory for each programme and stores the tokens in each.